repositories

+++ app/src/main/java/com/alexeyyuditsky/weatherapp/findCity/domain/FoundCity.kt
@@ -2,6 +2,6 @@ package com.alexeyyuditsky.weatherapp.findCity.domain
 
 data class FoundCity(
     val name: String,
-    val latitude: Double,
-    val longitude: Double
+    val latitude: Float,
+    val longitude: Float
 )
