+++ app/src/main/java/com/alexeyyuditsky/weatherapp/weather/domain/WeatherInCity.kt
@@ -2,5 +2,5 @@ package com.alexeyyuditsky.weatherapp.weather.domain
 
 data class WeatherInCity(
     val cityName: String,
-    val temperature: Double,
+    val temperature: Float,
 )
